Review fix: Moved the provider related code to a new file, reverted IntentWatcher code to clean this PR
