--- papiea-sdk/python/__tests__/procedures/create_object_input.yml
@@ -1,4 +1,0 @@
-CreateObjectInput:
-    type: string
-    title: Input value of CreateObject function
-    description: Name of the object to be created
