--- papiea-sdk/python/__tests__/procedures/ensure_bucket_exists_input.yml
@@ -1,4 +1,0 @@
-EnsureBucketExistsInput:
-    type: string
-    title: Input value of EnsureBucketExists function
-    description: Name of the bucket to be created/checked for existence
