--- papiea-sdk/python/__tests__/procedures/link_object_input.yml
@@ -1,14 +1,0 @@
-LinkObjectInput:
-    type: object
-    title: Input value of Link function
-    description: Information for the new object to be linked
-    required:
-        - object_name
-        - object_uuid
-    properties:
-        object_name:
-            type: string
-            description: Name of the new object to be linked 
-        object_uuid:
-            type: string
-            description: UUID of the object to link to
