--- papiea-sdk/python/__tests__/procedures/unlink_object_input.yml
@@ -1,14 +1,0 @@
-UnlinkObjectInput:
-    type: object
-    title: Input value of Unlink function
-    description: Information for the object to be unlinked
-    required:
-        - object_name
-        - object_uuid
-    properties:
-        object_name:
-            type: string
-            description: Name of the object to be unlinked 
-        object_uuid:
-            type: string
-            description: UUID of the object to unlink from
