+++ papiea-sdk/python/e2e_tests/kinds/object_kind.yml
@@ -31,4 +31,4 @@ object:
                     object_name:
                         type: string
                         description: Name of the object which refers to this object
-                    
+                    # bucket reference added in python code
