+++ papiea-sdk/python/e2e_tests/security/metadata_extension.yml
