+++ papiea-sdk/python/papiea/core.py
@@ -24,7 +24,6 @@ EntityReference = AttributeDict
 Spec = AttributeDict
 EntitySpec = AttributeDict
 Metadata = AttributeDict
-IntentWatcher = AttributeDict
 
 class Action(str):
     Read = "read"
