+++ papiea-sdk/python/papiea/python_sdk.py
@@ -10,7 +10,6 @@ from .core import (
     Entity,
     IntentfulExecutionStrategy,
     IntentfulSignature,
-    IntentWatcher,
     Kind,
     ProceduralExecutionStrategy,
     ProceduralSignature,
@@ -22,7 +21,7 @@ from .core import (
     Version, ProcedureDescription,
 )
 from .python_sdk_context import IntentfulCtx, ProceduralCtx
-from .python_sdk_exceptions import ApiException, InvocationError, PapieaBaseException, SecurityApiError
+from .python_sdk_exceptions import InvocationError, SecurityApiError
 from .utils import json_loads_attrs, validate_error_codes
 
 
@@ -127,69 +126,6 @@ class SecurityApi(object):
             return res
         except Exception as e:
             raise SecurityApiError.from_error(e, "Cannot deactivate s2s key")
-
-class IntentWatcherApi(object):
-    def __init__(
-        self,
-        papiea_url: str,
-        s2skey: Secret = None,
-        logger: logging.Logger = logging.getLogger(__name__)
-    ):
-        headers = {
-            "Content-Type": "application/json",
-        }
-
-        if s2skey is not None:
-            headers["Authorization"] = f"Bearer {s2skey}"
-        self.api_instance = ApiInstance(
-            f"{papiea_url}/services/intent_watcher", headers=headers, logger=logger
-        )
-
-        self.logger = logger
-
-    async def __aenter__(self) -> "IntentWatcherApi":
-        return self
-
-    async def __aexit__(
-        self,
-        exc_type: Optional[Type[BaseException]],
-        exc_val: Optional[BaseException],
-        exc_tb: Optional[TracebackType]
-    ) -> None:
-        await self.api_instance.close()
-
-    async def get_intent_watcher(self, id: str) -> IntentWatcher:
-        try:
-            return await self.api_instance.get(id)
-        except PapieaBaseException as papiea_exception:
-            raise Exception('Papiea exception: ' + str(papiea_exception))
-        except ApiException as api_exception:
-            raise Exception('Api exception: ' + api_exception.reason)
-        except Exception as ex:
-            raise ex
-
-    async def list_intent_watcher(self) -> List[IntentWatcher]:
-        try:
-            res = await self.api_instance.get("")
-            return res.results
-        except PapieaBaseException as papiea_exception:
-            raise Exception('Papiea exception: ' + str(papiea_exception))
-        except ApiException as api_exception:
-            raise Exception('Api exception: ' + api_exception.reason)
-        except Exception as ex:
-            raise ex
-
-    # filter_intent_watcher(AttributeDict(status='Pending'))
-    async def filter_intent_watcher(self, filter_obj: Any) -> List[IntentWatcher]:
-        try:
-            res = await self.api_instance.post("filter", filter_obj)
-            return res.results
-        except PapieaBaseException as papiea_exception:
-            raise Exception('Papiea exception: ' + str(papiea_exception))
-        except ApiException as api_exception:
-            raise Exception('Api exception: ' + api_exception.reason)
-        except Exception as ex:
-            raise ex
 
 class ProviderSdk(object):
     def __init__(
