+++ papiea-sdk/python/papiea/utils.py
@@ -17,19 +17,3 @@ def validate_error_codes(error_schemas: Optional[ErrorSchemas]):
             numeric_code = int(code)
             if not isinstance(numeric_code, int) or not (400 < numeric_code < 599):
                 raise Exception("Error description should feature status code in 4xx or 5xx")
-
-def ref_type(kind: str, description: str = '') -> AttributeDict:
-    return AttributeDict(
-        type='object',
-        description=description,
-        required=['uuid', 'kind'],
-        properties=AttributeDict(
-            uuid=AttributeDict(
-                type='string'
-            ),
-            kind=AttributeDict(
-                type='string',
-                example=kind or 'kind_name'
-            )
-        )
-    )
